refactor(menu)!: extract action execution logic to ActionDispacher
