+++ src/actions.rs
@@ -8,7 +8,8 @@ use std::process::Command;
 
 use crate::cli::{Args, Commands};
 use crate::menu::Menu;
-use crate::menu::event_handler::DefaultEventHanlder;
+use crate::menu::action_dispacher::DefaultActionDispacher;
+use crate::menu::event_handler::DefaultEventHandler;
 use crate::menu::item::MenuItem;
 use crate::menu::renderer::DefaultMenuRenderer;
 use crate::persistence::*;
@@ -196,7 +197,8 @@ fn menu(show_preview: bool, ask_for_confirmation: bool) -> Result<()> {
         show_preview,
         ask_for_confirmation,
         Box::new(DefaultMenuRenderer),
-        Box::new(DefaultEventHanlder),
+        Box::new(DefaultEventHandler),
+        Box::new(DefaultActionDispacher),
     );
 
     menu.run(&mut terminal)?;
